Assignment A2Q1 A2Q2 A2Q3 A2Q4(TEMP) A2Q5 A2Q6 A2Q7 A2Q8 A2Q9 Assignment A3Q1 A3Q2(NO WORK) A3Q3 A3Q4 Start

+++ src/Assignment3/A3Q3.java
@@ -18,7 +18,7 @@ public class A3Q3 {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
                 System.out.println("Please enter a number between 1 and 10.");
-                double Numba = input.nextInt();
+                int Numba = input.nextInt();
                 while(1>Numba || Numba>10){
                     System.out.println("Between 1 and 10.");
                     Numba = input.nextInt();
